+++ src/main/java/com/syzc/sseip/controller/DiseaseTypeController.java
@@ -1,10 +1,13 @@
 package com.syzc.sseip.controller;
 
 import com.syzc.sseip.entity.DiseaseType;
+import com.syzc.sseip.entity.User;
+import com.syzc.sseip.entity.enumtype.Role;
 import com.syzc.sseip.service.DiseaseTypeService;
 import com.syzc.sseip.util.HosException;
 import com.syzc.sseip.util.NotExistException;
 import com.syzc.sseip.util.Page;
+import com.syzc.sseip.util.exception.AuthException;
 import org.apache.log4j.Level;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -38,6 +41,11 @@ public class DiseaseTypeController {
 
     @RequestMapping(value = "/remove", method = RequestMethod.POST)
     public String remove(Long id, Model model, HttpServletRequest request, HttpSession session) {
+        User user = (User) session.getAttribute("loginUser");
+        if (user.getRole() != Role.ADMIN) {
+            throw AuthException.create("没有权限", Level.DEBUG);
+        }
+
         //删除操作
         Boolean r = diseaseTypeService.remove(id);
         if (!r) {
@@ -52,11 +60,21 @@ public class DiseaseTypeController {
 
     @RequestMapping(value = "/add", method = RequestMethod.GET)
     public String add(Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loginUser");
+        if (user.getRole() != Role.ADMIN) {
+            throw AuthException.create("没有权限", Level.DEBUG);
+        }
+
         return "disease-type-add";
     }
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
     public String add(DiseaseType diseaseType, Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loginUser");
+        if (user.getRole() != Role.ADMIN) {
+            throw AuthException.create("没有权限", Level.DEBUG);
+        }
+
         if (diseaseTypeService.add(diseaseType) != null) {
             model.addAttribute("diseaseType", diseaseType);
             return "disease-type-added";
@@ -67,6 +85,11 @@ public class DiseaseTypeController {
 
     @RequestMapping(value = "/update/{id:\\d+}", method = RequestMethod.GET)
     public String update(@PathVariable("id") Long id, Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loginUser");
+        if (user.getRole() != Role.ADMIN) {
+            throw AuthException.create("没有权限", Level.DEBUG);
+        }
+
         DiseaseType diseaseType = diseaseTypeService.get(id);
         if (diseaseType == null) {
             throw NotExistException.create(Level.DEBUG);
@@ -77,6 +100,11 @@ public class DiseaseTypeController {
 
     @RequestMapping(value = "/update/{id:\\d+}", method = RequestMethod.POST)
     public String update(@PathVariable("id") Long id, DiseaseType diseaseType, Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loginUser");
+        if (user.getRole() != Role.ADMIN) {
+            throw AuthException.create("没有权限", Level.DEBUG);
+        }
+
         diseaseType.setId(id);
         if (!diseaseTypeService.update(diseaseType)) {
             throw HosException.create("目标不存在", Level.DEBUG);
